fix(cli-1994): split the two files CLI-2004 and this change both grew
Rebasing onto develop brought the screen-reader mode into the same two files this change edits, and
both went over their size budgets: the workspace switcher past the hard 300-line limit, and App.tsx
past the baseline the screen-reader work had just tightened. The size rule says split, not grow, so
both are split along seams that were already there.

The switcher's selection state — where the cursor is, which slice is on screen, what a keypress does
to both — moves to its own module. The component keeps what it draws.

App.tsx gains one call instead of three: the background panel's detail read, its attach handler and
the selection derivation that only ever fed the read now sit together in `useBackgroundPanel`, since
they are one surface's data and App was keeping their inputs in step by hand.

The two branches had also each extracted the detail read into a hook of their own — CLI-2004's
`useExecutionDetailPage` and this change's `useExecutionWorkspaceDetail`, the same 69 lines under two
names. The one already on develop is kept and the duplicate is deleted.

Verified after: agent-framework 1680, agent-command 308, agent-executor 117, agent-transport-tui 797,
agent-subagent-runner 70, agent-cli 500 — all green; tsc clean; file-size passes with both files
re-frozen at their new, smaller numbers; ESLint four under the ceiling.

Lane: L2


Work-Run: 4b5027f7-d172-4828-b98d-9ab228e51726
Work-Receipt: g0-r0
