+++ packages/agent-transport-tui/src/ExecutionWorkspaceSwitcher.tsx
@@ -1,15 +1,9 @@
 import { Box, useInput } from 'ink';
-import React, { useEffect, useRef, useState } from 'react';
+import React from 'react';
 
+import { useWorkspaceSwitcherSelection } from './execution-workspace-switcher-selection.js';
 import { formatExecutionWorkspaceEntryRow } from './execution-workspace-view-model.js';
-import {
-  applySelectionInput,
-  createSelectionFlowState,
-  getVerticalSelectionInputAction,
-  normalizeSelectionState,
-  type ISelectionFlowState,
-  type TSelectionInputAction,
-} from './flows/selection-flow.js';
+import { getVerticalSelectionInputAction } from './flows/selection-flow.js';
 import { useNumberedSelection } from './hooks/useNumberedSelection.js';
 import {
   KeyHintFooter,
@@ -26,8 +20,6 @@ import type {
   IExecutionWorkspaceEntry,
   IExecutionWorkspaceSnapshot,
 } from '@robota-sdk/agent-interface-execution';
-
-const MAX_VISIBLE_WORKSPACE_ENTRIES = 8;
 
 /** Footer for the workspace switcher (order: navigate → primary → dismiss). */
 export const EXECUTION_WORKSPACE_SWITCHER_FOOTER_HINTS: readonly IKeyHint[] = [
@@ -96,7 +88,13 @@ export default function ExecutionWorkspaceSwitcher({
     (input, key) => {
       // CLI-1994: `a` attaches to the focused fork. Guarded against the modifiers because Ink
       // reports Ctrl+A as the letter with `ctrl` set, and a chord must not attach silently.
-      if (canAttach && input === ATTACH_KEY && key.ctrl !== true && key.meta !== true && focusedEntry) {
+      if (
+        canAttach &&
+        input === ATTACH_KEY &&
+        key.ctrl !== true &&
+        key.meta !== true &&
+        focusedEntry
+      ) {
         // Attaching replaces what the terminal is looking at, so the switcher has nothing left to
         // switch between — it closes itself rather than making every caller remember to.
         onAttach(focusedEntry);
@@ -160,100 +158,6 @@ export default function ExecutionWorkspaceSwitcher({
         />
       )}
     </Box>
-  );
-}
-
-interface IUseWorkspaceSwitcherSelectionInput {
-  entries: IExecutionWorkspaceEntry[];
-  selectedEntryId?: string;
-  onSelect: (entryId: string) => void;
-  onClose: () => void;
-}
-
-function useWorkspaceSwitcherSelection({
-  entries,
-  selectedEntryId,
-  onSelect,
-  onClose,
-}: IUseWorkspaceSwitcherSelectionInput): {
-  normalized: ISelectionFlowState;
-  visibleEntries: IExecutionWorkspaceEntry[];
-  applyAction: (action: TSelectionInputAction) => void;
-} {
-  const [state, setState] = useState<ISelectionFlowState>(() => createSelectionFlowState());
-  const stateRef = useRef(state);
-
-  useEffect(() => {
-    const selectedIndex = Math.max(
-      0,
-      entries.findIndex((entry) => entry.id === selectedEntryId),
-    );
-    const nextState = createNormalizedSelection({ selectedIndex, itemCount: entries.length });
-    stateRef.current = nextState;
-    setState(nextState);
-  }, [entries.length, selectedEntryId]);
-
-  const normalized = createNormalizedSelection({
-    selectedIndex: state.selectedIndex,
-    scrollOffset: state.scrollOffset,
-    itemCount: entries.length,
-  });
-  if (normalized !== state) stateRef.current = normalized;
-  return {
-    normalized,
-    visibleEntries: entries.slice(
-      normalized.scrollOffset,
-      normalized.scrollOffset + MAX_VISIBLE_WORKSPACE_ENTRIES,
-    ),
-    applyAction: createApplyAction({ entries, stateRef, setState, onSelect, onClose }),
-  };
-}
-
-function createApplyAction({
-  entries,
-  stateRef,
-  setState,
-  onSelect,
-  onClose,
-}: {
-  entries: IExecutionWorkspaceEntry[];
-  stateRef: React.MutableRefObject<ISelectionFlowState>;
-  setState: React.Dispatch<React.SetStateAction<ISelectionFlowState>>;
-  onSelect: (entryId: string) => void;
-  onClose: () => void;
-}): (action: TSelectionInputAction) => void {
-  return (action): void => {
-    const result = applySelectionInput(stateRef.current, action, {
-      itemCount: entries.length,
-      maxVisible: MAX_VISIBLE_WORKSPACE_ENTRIES,
-    });
-    const nextState =
-      result.effect.type === 'select' || result.effect.type === 'cancel'
-        ? { ...result.state, resolved: false }
-        : result.state;
-    stateRef.current = nextState;
-    setState(nextState);
-    if (result.effect.type === 'cancel') {
-      onClose();
-    } else if (result.effect.type === 'select') {
-      const entry = entries[result.effect.index];
-      if (entry) onSelect(entry.id);
-    }
-  };
-}
-
-function createNormalizedSelection(input: {
-  selectedIndex: number;
-  scrollOffset?: number;
-  itemCount: number;
-}): ISelectionFlowState {
-  return normalizeSelectionState(
-    {
-      selectedIndex: input.selectedIndex,
-      scrollOffset: input.scrollOffset ?? 0,
-      resolved: false,
-    },
-    { itemCount: input.itemCount, maxVisible: MAX_VISIBLE_WORKSPACE_ENTRIES },
   );
 }
