+++ scripts/harness/file-size-baseline.json
@@ -23,7 +23,7 @@
   "packages/agent-framework/src/interactive/interactive-session-execution-controller.ts": 412,
   "packages/agent-framework/src/interactive/interactive-session-history-tracker.ts": 359,
   "packages/agent-framework/src/interactive/interactive-session-skill-router.ts": 341,
-  "packages/agent-framework/src/interactive/interactive-session.ts": 1017,
+  "packages/agent-framework/src/interactive/interactive-session.ts": 987,
   "packages/agent-framework/src/testing/scripted-session-harness.ts": 454,
   "packages/agent-framework/src/transport-host/transport-registry.ts": 306,
   "packages/agent-framework/src/user-local/memory.ts": 310,
@@ -40,7 +40,7 @@
   "packages/agent-provider-openai-compatible/src/qwen/responses-parser.ts": 302,
   "packages/agent-session/src/session-run.ts": 322,
   "packages/agent-session/src/session.ts": 312,
-  "packages/agent-transport-tui/src/App.tsx": 571,
+  "packages/agent-transport-tui/src/App.tsx": 570,
   "packages/agent-transport-tui/src/TuiInteractionChannel.ts": 615,
   "packages/agent-transport-webrtc-web/src/client/rtc-session-client.ts": 399,
   "packages/agent-transport-ws/src/ws-transport-configurable.ts": 309,
