+++ scripts/harness/reference-kind-baseline.json
@@ -59,6 +59,7 @@
   ".agents/skills/vitest-testing-strategy/SKILL.md": 2,
   ".agents/spec-docs/active/AGREEMENT-008-coordinate-the-first-bulk-child-issue-absorption-wave.md": 1,
   ".agents/spec-docs/active/CLI-1990-deferred-tool-schemas-and-tool-search.md": 19,
+  ".agents/spec-docs/active/CLI-1994-fork-the-conversation-into-a-background-session.md": 10,
   ".agents/spec-docs/active/CLI-2004-tui-screen-reader-mode.md": 11,
   ".agents/spec-docs/active/INFRA-104-promotion-carries-closing-keywords-to-main.md": 10,
   ".agents/spec-docs/done/ARCH-029-command-host-capability-roles.md": 4,
